+++ pyctr/type/sdtitle.py
@@ -10,6 +10,10 @@ from pathlib import Path, PurePosixPath
 from typing import TYPE_CHECKING
 from weakref import WeakSet
 
+from fs import open_fs
+from fs.osfs import OSFS
+from fs.path import dirname as fs_dirname, join as fs_join
+
 from ..common import PyCTRError
 from ..crypto import add_seed
 from .ncch import NCCHReader
@@ -19,7 +23,9 @@ if TYPE_CHECKING:
     from pathlib import PurePath
     from typing import BinaryIO, Dict, List, Set, Union
 
-    from ..common import FilePath
+    from fs.base import FS
+
+    from ..common import FilePath, FilePathOrFS
     from .ncch import NCCHReader
     from .sd import SDFilesystem
     from .tmd import ContentChunkRecord
@@ -54,18 +60,19 @@ class SDTitleReader:
     Only NCCH contents are supported. SRL (DSiWare) contents are currently ignored.
 
     :param file: A path to a tmd file. All the contents should be in the same directory.
+    :param fs: An :meth:`FS` object or an `FS URL <https://docs.pyfilesystem.org/en/latest/openers.html>`.
     :param case_insensitive: Use case-insensitive paths for the RomFS of each NCCH container.
     :param dev: Use devunit keys.
     :param seed: Seed to use. This is a quick way to add a seed using :func:`~.seeddb.add_seed`.
     :param load_contents: Load each partition with :class:`~.NCCHReader`.
     :param sdfs: :class:`~.SDFilesystem` object to use, if opening contents that are currently encrypted. Usually this
-        should not be set directly, instead open the title through :class:`~.SDFilesystem`. (NYI)
+        should not be set directly, instead open the title through :class:`~.SDFilesystem`.
     :param sd_id1: ID1 to use if opened through :class:`~.SDFilesystem`.
     """
 
     __slots__ = (
         '_base_files', '_open_files', 'available_sections', 'closed', 'content_info', 'contents', 'sd_id1', 'sdfs',
-        'tmd'
+        'tmd', 'fs'
     )
 
     available_sections: 'List[Union[SDTitleSection, int]]'
@@ -86,8 +93,8 @@ class SDTitleReader:
     tmd: TitleMetadataReader
     """The :class:`~.TitleMetadataReader` object with information from the TMD section."""
 
-    def __init__(self, file: 'FilePath', *, case_insensitive: bool = False, dev: bool = False, seed: bytes = None,
-                 load_contents: bool = True, sdfs: 'SDFilesystem' = None, sd_id1: str = None):
+    def __init__(self, file: 'FilePath', *, fs: 'FS' = None, case_insensitive: bool = False, dev: bool = False,
+                 seed: bytes = None, load_contents: bool = True, sdfs: 'SDFilesystem' = None, sd_id1: str = None):
         self.closed = False
 
         self.sdfs = sdfs
@@ -107,12 +114,24 @@ class SDTitleReader:
         self.available_sections = []
 
         if self.sdfs:
+            # custom FS is not supported here (this section is deprecated anyway)
             file = PurePosixPath(file)
+            title_root = fsdecode(file.parent)
+            fs = OSFS(title_root)
         else:
-            file = Path(fsdecode(file)).absolute()
-        title_root = file.parent
+            if fs:
+                title_root = fs_dirname(file)
+            else:
+                file = Path(file).absolute()
+                title_root = fsdecode(file.parent)
+                fs = OSFS(title_root)
+        self.fs = fs
 
-        def add_file(section: 'Union[SDTitleSection, int]', path: "PurePath"):
+        print(title_root)
+        print(fs)
+        print(fs.listdir(title_root))
+
+        def add_file(section: 'Union[SDTitleSection, int]', path: str):
             self._base_files[section] = path
             self.available_sections.append(section)
 
@@ -130,13 +149,13 @@ class SDTitleReader:
 
             # this should ideally never be uppercase in practice
             # since the console stores these as lowercase
-            content_file = title_root / (record.id + '.app')
+            content_file = fs_join(title_root, record.id + '.app')
             if self.sdfs:
                 if not self.sdfs.isfile(str(content_file), id1=self.sd_id1):
                     # can't find the file, so continue to the next record
                     continue
             else:
-                if not content_file.is_file():
+                if not fs.isfile(content_file):
                     continue
 
             self.content_info.append(record)
@@ -190,6 +209,6 @@ class SDTitleReader:
         if self.sdfs:
             f = self.sdfs.open(str(filepath), 'rb', id1=self.sd_id1)
         else:
-            f = open(filepath, 'rb')
+            f = self.fs.open(filepath, 'rb')
         self._open_files.add(f)
         return f